# vts_server/app/api/api.yaml
swagger: "2.0"
info: 
  title: GMS APIs
  description: VTS GMS api list
  version: 1.0 
  contact: 
    email: ykim5470@vreducation.kr
  servers: 
    -url: "https://live.enjoystreet.kr"
    description: Server host url 
tags: 
- name: "Auth"
  description: "로그인"
- name: "Live Management"
  description: "라이브관리"
- name: "My Page"
  description: "마이페이지"
- name: "Live Tour (Overall)"
  descriptionn: "라이브투어(일반)"

schemes: [https]

paths:
  /guideLogin: 
    post:
      tags:
      - "Auth"
      summary: "Guide User Sign in to the platform"
      description: ""
      consumes: 
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "email and password are required to log in"
        required: true 
        schema:
          type: object
          properties:
            email:
              type: string
            password:
              type: string
          example: {email: ykim5470@gmail.com, password: password}
      responses:
        405:
          description: "Invalid input Request"
        200:
          description: "Success"
        404:
          description: "Not Found Error"
        500:
          description: "Internal Server Error"
  /createRoomNumber:
    get:
      tags:
      - "Live Management"
      summary: "Get a random room number"
      description: ""
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            type: object
            properties:
              roomId: 
                type: string
                pattern: '^₩d{3}-₩d{2}-₩d{4}$'
            examples: {roomId: "1-7d3a5976faed4f278f7bfd9c0f56b55d"}
        500:
          description: "Internal Server Error"
  /brandList:
    get:
      tags:
      - "Live Management"
      summary: "Get registered brand list"
      description: ""
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            type: array
            items:
              type: object
              properties:
                Id:
                  type: integer
                Name:
                  type: string
              examples: [{Id: 1, Name: DummyStore1}]
        500:
          description: "Internal Server Error" 
  /roomCreate:
      post:
        tags:
        - "Live Management"
        summary: "uploads an image"
        description: ""
        operationId: "uploadFile"
        consumes:
        - "multipart/form-data"
        produces:
        - "application/json"
        parameters:
        - name: "guideInfo"
          in: "formData"
          required: true
          type: object
        - name: "title"
          in: "formData"
          type: string
        - name: "roomId"
          in: "formData"
          type: string
        - name: "brandConfig"
          in: "formData"
          type: object
        - name: "thumbnail"
          in: "formData"
          description: "file to upload"
          required: true
          type: file
        responses:
          200:
            description: Success
          400:
            description: User Error (Invalid Request)
          500:
            description: Internal Server Error 


  /recordMediaUpload:
      post:
        tags:
        - "Live Management"
        summary: "uploads an record media thumbnail and media file"
        description: ""
        operationId: "uploadFile"
        consumes:
        - "multipart/form-data"
        produces:
        - "application/json"
        parameters:
        - name: "roomId"
          in: "formData"
          required: true
          type: string
        - name: "title"
          in: "formData"
          type: string
        - name: "host"
          in: "formData"
          type: integer
        - name: "brandConfig"
          in: "formData"
          type: object
        - name: "resources"
          in: "formData"
          description: "Image file to upload"
          required: true
          type: file
        - name: "resources"
          in: "formData"
          description: "Media file to upload"
          required: true
          type: file
        responses:
          200:
            description: Success
          400:
            description: User Error (Invalid Request)
          500:
            description: Internal Server Error 


  